Compile Error Fixes
Merged two 'actor' scripts into one.
