--- Assets/StoryDialogue/statTestSheep.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2bd1776e532453944acff84610a8f5e4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
